Implemented exchange rates and viewing the necessary figures for buying, selling, and exchanging.

+++ src/model/FiatCurrency.java
@@ -2,9 +2,9 @@ package model;
 
 public class FiatCurrency extends Currency {
     private String sign;
-    private float subunit;
+    private double subunit;
 
-    public FiatCurrency(String name, String code, String sign, float subunit) {
+    public FiatCurrency(String name, String code, String sign, double subunit) {
         super(name, code);
         this.sign = sign;
         this.subunit = subunit;
